remove redundancy code
