--- manager/pkg/cronjob/scheduler_suite_test.go
@@ -1,94 +1,0 @@
-// Copyright (c) 2022 Red Hat, Inc.
-// Copyright Contributors to the Open Cluster Management project
-
-package cronjob_test
-
-import (
-	"context"
-	"path/filepath"
-	"testing"
-	"time"
-
-	"github.com/jackc/pgx/v4/pgxpool"
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-	"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/client-go/rest"
-	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/envtest"
-	logf "sigs.k8s.io/controller-runtime/pkg/log"
-	"sigs.k8s.io/controller-runtime/pkg/log/zap"
-
-	"github.com/stolostron/multicluster-global-hub/manager/pkg/cronjob"
-	// managerscheme "github.com/stolostron/multicluster-global-hub/manager/pkg/scheme"
-	"github.com/stolostron/multicluster-global-hub/pkg/database"
-	"github.com/stolostron/multicluster-global-hub/test/pkg/testpostgres"
-)
-
-var (
-	testenv      *envtest.Environment
-	cfg          *rest.Config
-	ctx          context.Context
-	cancel       context.CancelFunc
-	mgr          ctrl.Manager
-	pool         *pgxpool.Pool
-	testPostgres *testpostgres.TestPostgres
-)
-
-func TestScheduler(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Scheduler Suite")
-}
-
-var _ = BeforeSuite(func() {
-	logf.SetLogger(zap.New(zap.WriteTo(GinkgoWriter), zap.UseDevMode(true)))
-	ctx, cancel = context.WithCancel(context.Background())
-
-	By("Prepare envtest environment")
-	var err error
-	testenv = &envtest.Environment{
-		CRDDirectoryPaths: []string{
-			filepath.Join("..", "..", "..", "pkg", "testdata", "crds"),
-		},
-		ErrorIfCRDPathMissing: true,
-	}
-	cfg, err = testenv.Start()
-	Expect(err).NotTo(HaveOccurred())
-	Expect(cfg).NotTo(BeNil())
-
-	By("Create test postgres")
-	testPostgres, err = testpostgres.NewTestPostgres()
-	Expect(err).NotTo(HaveOccurred())
-
-	pool, err = database.PostgresConnPool(ctx, testPostgres.URI, "ca-cert-path")
-	Expect(err).NotTo(HaveOccurred())
-
-	mgr, err = ctrl.NewManager(cfg, ctrl.Options{
-		MetricsBindAddress: "0",
-		Scheme:             scheme.Scheme,
-	})
-	Expect(err).NotTo(HaveOccurred())
-
-	// By("Add to Scheme")
-	// Expect(managerscheme.AddToScheme(mgr.GetScheme())).NotTo(HaveOccurred())
-	Expect(cronjob.AddSchedulerToManager(context.TODO(), mgr, pool, "month", false)).Should(Succeed())
-	Expect(cronjob.AddSchedulerToManager(context.TODO(), mgr, pool, "week", false)).Should(Succeed())
-	Expect(cronjob.AddSchedulerToManager(context.TODO(), mgr, pool, "day", false)).Should(Succeed())
-	Expect(cronjob.AddSchedulerToManager(context.TODO(), mgr, pool, "hour", false)).Should(Succeed())
-	Expect(cronjob.AddSchedulerToManager(context.TODO(), mgr, pool, "minute", false)).Should(Succeed())
-	Expect(cronjob.AddSchedulerToManager(context.TODO(), mgr, pool, "second", false)).Should(Succeed())
-})
-
-var _ = AfterSuite(func() {
-	cancel()
-	Expect(testPostgres.Stop()).NotTo(HaveOccurred())
-
-	By("Tearing down the test environment")
-	err := testenv.Stop()
-	// https://github.com/kubernetes-sigs/controller-runtime/issues/1571
-	// Set 4 with random
-	if err != nil {
-		time.Sleep(4 * time.Second)
-	}
-	Expect(testenv.Stop()).NotTo(HaveOccurred())
-})
