+++ manager/pkg/cronjob/scheduler_test.go
@@ -1,174 +1,72 @@
-// Copyright (c) 2022 Red Hat, Inc.
-// Copyright Contributors to the Open Cluster Management project
-
-package cronjob_test
+package cronjob
 
 import (
-	"fmt"
+	"context"
+	"path/filepath"
+	"testing"
 	"time"
 
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
+	"github.com/jackc/pgx/v4/pgxpool"
+	"github.com/stretchr/testify/assert"
+	"k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/client-go/rest"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/envtest"
+
+	"github.com/stolostron/multicluster-global-hub/pkg/database"
+	"github.com/stolostron/multicluster-global-hub/test/pkg/testpostgres"
 )
 
-var _ = Describe("sync the compliance data", Ordered, func() {
-	const targetSchema = "local_status"
-	const sourceTable = "compliance"
-	const targetTable = "compliance_history"
+var (
+	testenv      *envtest.Environment
+	cfg          *rest.Config
+	ctx          context.Context
+	cancel       context.CancelFunc
+	mgr          ctrl.Manager
+	pool         *pgxpool.Pool
+	testPostgres *testpostgres.TestPostgres
+)
 
-	BeforeAll(func() {
-		By("Creating test table in the database")
-		conn, err := pool.Acquire(ctx)
-		Expect(err).ToNot(HaveOccurred())
-		defer conn.Release()
-		_, err = conn.Exec(ctx, `
-			CREATE SCHEMA IF NOT EXISTS local_status;
-			CREATE SCHEMA IF NOT EXISTS status;
-			DO $$ BEGIN
-				CREATE TYPE local_status.compliance_type AS ENUM (
-					'compliant',
-					'non_compliant',
-					'unknown'
-				);
-			EXCEPTION
-				WHEN duplicate_object THEN null;
-			END $$;
-			DO $$ BEGIN
-				CREATE TYPE status.error_type AS ENUM (
-					'disconnected',
-					'none'
-				);
-			EXCEPTION
-				WHEN duplicate_object THEN null;
-			END $$;
-			CREATE TABLE IF NOT EXISTS local_status.compliance (
-				id uuid NOT NULL,
-				cluster_name character varying(63) NOT NULL,
-				leaf_hub_name character varying(63) NOT NULL,
-				error status.error_type NOT NULL,
-				compliance local_status.compliance_type NOT NULL,
-				cluster_id uuid
-			);
-			CREATE TABLE IF NOT EXISTS local_status.compliance_history (
-				id uuid NOT NULL,
-				cluster_id uuid NOT NULL,
-				updated_at timestamp without time zone DEFAULT now() NOT NULL,
-				compliance local_status.compliance_type NOT NULL,
-				compliance_changed_frequency integer NOT NULL DEFAULT 0
-			);
-			CREATE TABLE IF NOT EXISTS local_status.compliance_history_job_log (
-				name varchar(63) NOT NULL,
-				start_at timestamp NOT NULL DEFAULT now(),
-				end_at timestamp NOT NULL DEFAULT now(),
-				total int8,
-				inserted int8,
-				offsets int8, 
-				error TEXT
-			);`)
-		Expect(err).ToNot(HaveOccurred())
-		By("Check whether the tables are created")
-		Eventually(func() error {
-			rows, err := conn.Query(ctx, "SELECT * FROM pg_tables")
-			if err != nil {
-				return err
-			}
-			defer rows.Close()
-			expectedTableSet := map[string]bool{
-				sourceTable: true,
-				targetTable: true,
-			}
+func TestScheduler(t *testing.T) {
+	ctx, cancel = context.WithCancel(context.Background())
+	var err error
+	testenv = &envtest.Environment{
+		CRDDirectoryPaths: []string{
+			filepath.Join("..", "..", "..", "pkg", "testdata", "crds"),
+		},
+		ErrorIfCRDPathMissing: true,
+	}
+	cfg, err = testenv.Start()
+	assert.Nil(t, err)
+	assert.NotNil(t, cfg)
 
-			for rows.Next() {
-				columnValues, _ := rows.Values()
-				schema := columnValues[0]
-				table := columnValues[1]
-				if schema == targetSchema && expectedTableSet[table.(string)] {
-					delete(expectedTableSet, table.(string))
-				}
-			}
-			if len(expectedTableSet) > 0 {
-				return fmt.Errorf("tables %v are not created", expectedTableSet)
-			}
-			return nil
-		}, 10*time.Second, 2*time.Second).ShouldNot(HaveOccurred())
+	testPostgres, err = testpostgres.NewTestPostgres()
+	assert.Nil(t, err)
 
-		By("Start the manager")
-		go func() {
-			defer GinkgoRecover()
-			Expect(mgr.Start(ctx)).ToNot(HaveOccurred(), "failed to run manager")
-		}()
-		By("Waiting for the manager to be ready")
-		Expect(mgr.GetCache().WaitForCacheSync(ctx)).To(BeTrue())
+	pool, err = database.PostgresConnPool(ctx, testPostgres.URI, "ca-cert-path")
+	assert.Nil(t, err)
+
+	mgr, err = ctrl.NewManager(cfg, ctrl.Options{
+		MetricsBindAddress: "0",
+		Scheme:             scheme.Scheme,
 	})
+	assert.Nil(t, err)
 
-	It("sync the data from the source table to the target", func() {
-		By("Create the data to the source table")
-		_, err := pool.Exec(ctx, `
-			INSERT INTO local_status.compliance (id, cluster_name, leaf_hub_name, error, compliance, cluster_id) VALUES
-			('f8c4479f-fec5-44d8-8060-da9a92d5e138', 'local_cluster1', 'leaf1', 'none', 'compliant', 
-			'0cd723ab-4649-42e8-b8aa-aa094ccf06b4'),
-			('f8c4479f-fec5-44d8-8060-da9a92d5e139', 'local_cluster2', 'leaf2', 'none', 'compliant', 
-			'0cd723ab-4649-42e8-b8aa-aa094ccf06b4'),
-			('f8c4479f-fec5-44d8-8060-da9a92d5e137', 'local_cluster3', 'leaf3', 'none', 'compliant', 
-			'0cd723ab-4649-42e8-b8aa-aa094ccf06b4');
-		`)
-		Expect(err).ToNot(HaveOccurred())
+	assert.Nil(t, AddSchedulerToManager(context.TODO(), mgr, pool, "month", false))
+	assert.Nil(t, AddSchedulerToManager(context.TODO(), mgr, pool, "week", false))
+	assert.Nil(t, AddSchedulerToManager(context.TODO(), mgr, pool, "day", false))
+	assert.Nil(t, AddSchedulerToManager(context.TODO(), mgr, pool, "hour", false))
+	assert.Nil(t, AddSchedulerToManager(context.TODO(), mgr, pool, "minute", false))
+	assert.Nil(t, AddSchedulerToManager(context.TODO(), mgr, pool, "second", false))
 
-		By("Check whether the data is copied to the target table")
-		Eventually(func() error {
-			rows, err := pool.Query(ctx, "SELECT id, cluster_id, compliance FROM local_status.compliance_history")
-			if err != nil {
-				return err
-			}
-			defer rows.Close()
-			syncCount := 0
-			expectRecordMap := map[string]string{
-				"f8c4479f-fec5-44d8-8060-da9a92d5e138": "compliant",
-				"f8c4479f-fec5-44d8-8060-da9a92d5e139": "compliant",
-				"f8c4479f-fec5-44d8-8060-da9a92d5e137": "compliant",
-			}
-			for rows.Next() {
-				var id, cluster_id, compliance string
-				err := rows.Scan(&id, &cluster_id, &compliance)
-				if err != nil {
-					return err
-				}
-				fmt.Println("found record", "id", id, "cluster_id", cluster_id, "compliance", compliance)
-				if status, ok := expectRecordMap[id]; ok && status == compliance {
-					syncCount++
-					delete(expectRecordMap, id)
-				}
-			}
-			if syncCount != 3 {
-				return fmt.Errorf("table local_status.compliance_history records are not synced")
-			}
-			return nil
-		}, 20*time.Second, 2*time.Second).ShouldNot(HaveOccurred())
-
-		By("Check whether the job log is created")
-		Eventually(func() error {
-			rows, err := pool.Query(ctx, `SELECT name, total, inserted, offsets, error FROM 
-			local_status.compliance_history_job_log`)
-			if err != nil {
-				return err
-			}
-			defer rows.Close()
-
-			logCount := 0
-			for rows.Next() {
-				var name, errMessage string
-				var total, inserted, offsets int64
-				err := rows.Scan(&name, &total, &inserted, &offsets, &errMessage)
-				if err != nil {
-					return err
-				}
-				logCount += 1
-				fmt.Println("found log", "name", name, "total", total, "inserted", inserted, "offsets", offsets)
-			}
-			if logCount == 0 {
-				return fmt.Errorf("table local_status.compliance_history_job_log records are not synced")
-			}
-			return nil
-		}, 20*time.Second, 2*time.Second).ShouldNot(HaveOccurred())
-	})
-})
+	err = testPostgres.Stop()
+	assert.Nil(t, err)
+	cancel()
+	err = testenv.Stop()
+	// https://github.com/kubernetes-sigs/controller-runtime/issues/1571
+	// Set 4 with random
+	if err != nil {
+		time.Sleep(4 * time.Second)
+		assert.Nil(t, testenv.Stop())
+	}
+}
